+++ src/main/java/samdasu/jejuddai/repository/MenuRepository.java
@@ -5,5 +5,6 @@ import samdasu.jejuddai.entity.menu;
 
 import java.util.List;
 
-public interface MenuRepository extends JpaRepository<menu, Long> { List<menu> findByStoreId(Long storeId);
+public interface MenuRepository extends JpaRepository<menu, Long> {
+    List<menu> findByStoreId(Long storeId);
 }
